Fix bug in PublisherCrudService
Do database transaction

+++ service/publisherCrudService.js
@@ -16,30 +16,38 @@ exports.readPublishers = () => {
 
 exports.deletePublisher = (publisherId) => {
   const results = {
+    transactionError: false,
     readError: false,
     publisherNotFound: false,
     deleteError: false,
   };
   return new Promise((resolve, reject) => {
-    publisherDao.readPublishers(db, publisherId).then(
-      (readResult) => {
-        if (readResult.length === 0) {
-          results.publisherNotFound = true;
-          reject(results);
-          return;
-        }
-        publisherDao.deletePublisher(db, publisherId).then(
-          (deleteResult) => resolve(results),
-          (deleteError) => {
-            results.deleteError = true;
-            reject(results);
-          }
-        );
-      },
-      (readError) => {
-        results.readError = true;
+    db.beginTransaction((transactionError) => {
+      if (transactionError) {
+        results.transactionError = true;
         reject(results);
+        return;
       }
-    );
+      publisherDao.readPublishers(db, publisherId).then(
+        (readResult) => {
+          if (readResult.length === 0) {
+            results.publisherNotFound = true;
+            rollback(() => reject(results));
+            return;
+          }
+          publisherDao.deletePublisher(db, publisherId).then(
+            (deleteResult) => commit(() => resolve(results)),
+            (deleteError) => {
+              results.deleteError = true;
+              rollback(() => reject(results));
+            }
+          );
+        },
+        (readError) => {
+          results.readError = true;
+          rollback(() => reject(results));
+        }
+      );
+    });
   });
 };
